apps/backend/requirements.txt
# Runtime dependencies
fastapi
uvicorn[standard]
python-dotenv
pydantic[dotenv]
pydantic-settings
python-multipart   # future‑proof uploads
pytest
httpx  # for TestClient
supabase>=2.0.0  # official Python client
openai>=1.23  # for embeddings
sqlalchemy>=2.0.0  # for database ORM
psycopg2-binary  # PostgreSQL database adapter
python-jose[cryptography]  # for JWT tokens
passlib[bcrypt]  # for password hashing
python-multipart  # for file uploads
aiofiles  # for async file operations

# Event streaming
faust-streaming>=0.10.0
aiokafka>=0.8.0

# Development dependencies
pre-commit>=3.0.0
ruff>=0.4.1
black>=24.4.2
isort>=5.13.2
pytest-asyncio>=0.21.0